+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log : GitHub Repository Manager
 
+## 0.5.0 - Aug 04, 2020
+
+Added support to organizations repositories ([#10](https://github.com/SrBrahma/GitHub-Repository-Manager/pull/10))! You will need to re-OAuth again or add the 'org:read' to your Personal Acess Token permissions.
+
+Many thanks to [jonathan-fielding](https://github.com/jonathan-fielding) for this feature!
+
+
 ## 0.4.0 - Jul 08, 2020
 
 Added 'one click to clone' setting ([#7](https://github.com/SrBrahma/GitHub-Repository-Manager/pull/7))